Associate bug IDs with unimplemented functionality
Also correct uses of UNIMPLEMENTED() for functionality that is not
advertised as supported, and add spec quotes to justify and document
them.

This change also enforces requiring that the UNIMPLEMENTED() macro's
format string starts with a bug reference of the form b/###. Note we
still have to mention the bug ID in a FIXME() comment as well for
'go/todo' issue tracking (until b/148221114 gets fixed).

Bug: b/131243109
Change-Id: Ieb28e31f2c6c80a27b833626538e7e223b5fda08