Fix sRGB conversion precision issues
When sampling an sRGB texture format, we have a vector of 16-bit
elements where 0xFF00 represents 1.0, but we were using a lookup table
which produced results in a 0xFFFF base.

Also on linear to sRGB conversion, precision was lost due to truncating
to 12-bit. Adding a rounding factor before truncating preserves
precision.

Bug: b/149574741
Change-Id: I19218c958867c4b5e97dc3b62b4ea83f37161134