SpirvShaderDebugger: Expose builtins in HLSL / GLSL.
Previously, the builtins were only exposed in the SPIR-V scope.
This change utilizes the new `VariableContainer::extend()` functionality so that a single `VariableContainer` is created to hold all the builtins, and this is used as a base for the SPIR-V locals VC, as well as the high-level local scope VCs.

As part of this change, I've ditched the concept of non-per-lane builtins.
While some builtins are uniform across all lanes, having some at the top level of the watch tree, and others in the `Lane N` tree was just messy and complex to test.

Finally, a handful of builtins are also exposed for vertex shaders.

Bug: b/148401179