Update patch set 2
Patch Set 2: Presubmit-Ready+1

Ok only SampledImageArrayDynamicIndexing is problematic, and I wonder why. There is this code:

	if(getType(baseObject.type).storageClass == spv::StorageClassUniformConstant)
	{
		// indexing into an array of descriptors.
		auto &obj = getObject(indexIds[i]);
		if(obj.kind != Object::Kind::Constant)
		{
			UNSUPPORTED("SPIR-V SampledImageArrayDynamicIndexing Capability");
		}
		...
		ptr.base += stride * GetConstScalarInt(indexIds[i]);
	}

Which requires the index to be a constant. However, if the storage class is not UniformConstant (for example for storage buffers), the `else` part of the above if handles constants and non constants correctly:

	else
	{
		...
		auto &obj = getObject(indexIds[i]);
		if(obj.kind == Object::Kind::Constant)
		{
			ptr += stride * GetConstScalarInt(indexIds[i]);
		}
		else
		{
			ptr += SIMD::Int(stride) * state->getIntermediate(indexIds[i]).Int(0);
		}
	}

I'm going to try to see if using intermediate for UniformConstant works on the bots or not, but perhaps you have an idea why there is such a limitation?

It looks like since ptr.base is modified (instead of ptr), you expect the values to be in fact dynamically uniform, but "Kind" doesn't provide enough information to distinguish between "just any value" and "dynamically uniform values", so you resort to Constant alone. Is my understanding correct? Can we do something about it?

Patch-set: 2
Reviewer: Gerrit User 46365 <46365@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Presubmit-Ready=+1