Split JIT creation from LLVMReactor.cpp into LLVMJIT.cpp
This makes a rough distinction between code that mainly only interacts
with LLVM's IRBuilder interface, and the part that sets up the JIT and
manages module and function creation, memory management, and external
symbol resolution.

The latter much more often requires changes when LLVM is upgraded than
the former, so it's useful to keep them apart as we'll have multiple
versions of LLVM in different build environments.

Bug: b/148561024
Change-Id: I2ac2d00b4f3bcd277289a4ec6856c9eb7bb6aa24