Reactor: Copy System/Debug code to Reactor copy
This change keeps the implementationis separate, but at least they're now consistent.
System/Debug requires `UNIMPLEMENTED()` to have a bug number for each message, but reactor has a whole bunch of ancient uses without any mention of a bug. I don't have the context to create these bugs, so for now I've created a temporary `UNIMPLEMENTED_NO_BUG()` variant. With this we can easily grep the codebase to locate these usages and create bugs as needed.

Bug: b/146051794
Bug: b/144093703
Change-Id: Id5208f924e63dae7443f95aab5b601a290c297e7