Don't use std::initializer_list for array views.
The team debated whether this was legal or not at the time.
It seems that at least one of Android's compilers generates code that stomps the list data after the `std::initializer_list` is declared and before it is used.

Switch to using regular C arrays with constexpr / std::array.

Fixes: b/149235682
Change-Id: Ief95dc4c6b4aee0c181bbe34aff19aa95f5fac8c