Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > It's cool, but I think we should put this in a new "tools/" directory, where we can place random scripts to help with stuff like this. Otherwise we might end up with multiple scripts in third_party, and it can get messy fast.
> 
> What's the issue with having multiple scripts at the root third_party folder when they're specifically used for updating third_party code?

As I wrote, it can get messy fast. We may end up with scripts that call other scripts in there, and we'll end up with a bunch of scripts in the root. We might want to factor out some and stick 'em in a directory, and it would be weird for that to be in third_party.

Generally, I've seen these types of scripts put into a separate folder, like "tools". Could be in "tools/third_party/" for instance.

Patch-set: 1
Reviewer: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>