Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > It's cool, but I think we should put this in a new "tools/" directory, where we can place random scripts to help with stuff like this. Otherwise we might end up with multiple scripts in third_party, and it can get messy fast.
> > > 
> > > What's the issue with having multiple scripts at the root third_party folder when they're specifically used for updating third_party code?
> > 
> > As I wrote, it can get messy fast. We may end up with scripts that call other scripts in there, and we'll end up with a bunch of scripts in the root. We might want to factor out some and stick 'em in a directory, and it would be weird for that to be in third_party.
> > 
> > Generally, I've seen these types of scripts put into a separate folder, like "tools". Could be in "tools/third_party/" for instance.
> 
> EDIT: I very much like this document that vector-of-bool (a fairly well known C++ dude) put together a while back: https://api.csswg.org/bikeshed/?force=1&url=https://raw.githubusercontent.com/vector-of-bool/pitchfork/develop/data/spec.bs
> 
> He talks about the tools/ folder here:
> https://api.csswg.org/bikeshed/?force=1&url=https://raw.githubusercontent.com/vector-of-bool/pitchfork/develop/data/spec.bs#tld.tools

So, I actually considered adding putting this in a tools folder, but I had a vague memory of originally putting regres up for review in tools folder and being asked to place it in subdirectory off tests to keep the number of root folders down.

Anyway - I'm not convinced that this is going to get out of hand. We have one script used to update one third_party repo. Even if we had one .sh per project (which we won't), this doesn't seem unreasonable.

Also: this script doesn't actually do anything relative to third_party (it actually requires running from the SwiftShader root, which is perhaps something I should fix!), so relocating the script (when you prove me wrong!) is not difficult.

Patch-set: 1