Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> This information can be obtained from Gerrit.

It is obtaining the information from gerrit. 😊

> For example changes I reviewed in January:
> 
> https://swiftshader-review.googlesource.com/q/is:merged+reviewer:nicolascapens%2540google.com+after:2020-01-01+before:2020-01-31

Yes - the tool prints the query.

> It's 25 changes per page, but you can skip ahead in the URL if it's many.

You can, but who wants to do that?

> Would probably be a useful feature request for Gerrit to show the total.

This also figures out total lines added and removed, using gross (but effective) blacklisting. I don't expect the gerrit team are going to add this to their UI.

Again, I didn't assign reviewers as I wasn't wanting this to be reviewed. Just seemed like an easy way to share with the team.

Patch-set: 1