Update patch set 2
Patch Set 2: Code-Review-2

(1 comment)

This is not correct. A Vulkan implementation is not supposed to need to apply these rules itself, and *must not* assume that the layout is exactly GL's std140. Rather, conforming to the std140 rules implies conforming to the Vulkan rules, which are more flexible.

Patch-set: 2
Reviewer: Gerrit User 6982 <6982@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=-2