Log instead of warn of unsupported extensions
Currently we produce a warning every time an unsupported extension is
used. The original bug (b/139528538) called for logging these structs
silently unless a debugger's attached.

So I've replaced all of our warnings for unsupported structs with
calls to LOG_TRAP. This is an update to TRACE_ASSERT. LOG_TRAP will only
emit to a debug log if a Verbose variable is set to any level that isn't
Level::Disabled. "Disabled" a new level of logging that prevents anything
from being emitted. The Verbose variable is a global that can be set
during debugging or changed and recompiled. This variable also controls
whether logv prints to a file.

Bug: b/148415347
Change-Id: Ib4ad2b20b3dffce4fac597c891b2f5ee23e032c4