Log instead of warn of unsupported extensions
Currently we produce a warning every time an unsupported extension is
used. The original bug (b/139528538) called for logging these structs
silently unless a debugger's attached.

So I've replaced all of our warnings for unsupported structs with
calls to LOG_ASSERT. This is an update to TRACE_ASSERT that tells logv
in Debug.hpp to just Record the message when we're not attached to a
debugger. "Record" is a new level of logging that tells our logging
functions to avoid emitting any information to stdout or stderr. So info
is only logged when someone explicitly enables logging to a file inside
Debug.cpp and recompiles.

Bug: b/148415347
Change-Id: Ib4ad2b20b3dffce4fac597c891b2f5ee23e032c4