Replace target_cpu with current_cpu in Build.gn files
A recent change was made that failed when building nacl, because despite
the target_cpu being set to x86, current_cpu is x64 for nacl, as it is
cross-compiled to x64 using a different toolchain. As such, we were
building and linking the x86 version of Subzero with the x64 version of
nacl, resulting in a linker error.

See the following doc for more info:
https://chromium.googlesource.com/dart/dartium/src/+/refs/heads/master/tools/gn/docs/cross_compiles.md#as-a-build_gn-author

Bug: b/150008309