Update patch set 1
Patch Set 1: Code-Review+2

From our offline chat, I said:

I had a brief encounter with -Bsymbolic in the past, and it was all pain
there seem to be a good number of articles out there warning about its use, as it breaks ODR
https://software.intel.com/en-us/articles/performance-tools-for-software-developers-bsymbolic-can-cause-dangerous-side-effects

A few more dangers listed in the SO anwers here: https://stackoverflow.com/questions/7216973/is-there-a-downside-to-using-bsymbolic-functions

The top SO answer mentions making use of --dynamic-list to specifically name the symbols
maybe that would be better?


Nicolas replied:

Our ICD implement a C API which only exchanges POD and has a rigorous ABI, so many of the caveats for -Bsymbolic like exceptions and vtables don't apply. Also note that it is recommended by the Vulkan Loader documentation, and they use it for validation layers themselves. Also note -Bsymbolic only applies to the exported symbols, and those are already limited to those in the version script. So I think we'll be alright (famous last words).

Patch-set: 1
Reviewer: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Code-Review=+2