PipelineCache: Handle nullptr mismatch of specialization info.
It's valid in Vulkan for a shader to have specialization constants
and the pipeline to be lacking specialization constant data. This
could lead to a key match between the two even though there should
be a different shader payload. From the spec:

> It is legal for a SPIR-V module with specializations to be
> compiled into a pipeline where no specialization info was provided.
> SPIR-V specialization constants contain default values such that
> if a specialization is not provided.

Fix this by adding in the required checks. Should not affect perf.

Bug: angleproject:4426
Change-Id: I5f384420bdc1f03a36852db0b20393ae56452b62