Fix ASTC support for GN build
Defines in BUILD.gn files don't apply transitively to dependencies, so
we have to repeat the SWIFTSHADER_ENABLE_ASTC for the Device
subdirectory.

Note we could also use 'public_configs = [:swiftshader_libvulkan_private_config]'
but this adds more defines and build flags than what we need. Since this
is intended as a temporary workaround anyway, we're keeping it simple.

Bug: b/150600814