Subzero: fix bug introduced in Optimizer in b3d9a2adab8
When detecting and removing intermediate stores, if the storeValue is a
constant, we skip it; however, we must make sure to set unmatchedLoads
to true, or the optimizer will incorrectly eliminate the skipped store
on a later iteration.

This bug lead to incorrect generated code, which failed certain deqp-vk
tests.

Bug: b/148272103
Change-Id: I024d75decdc4f1d13e7f82cade50f8be09d68f3e