Explicitly compile with warnings-as-errors
This is a partial revert of https://swiftshader-review.googlesource.com/c/SwiftShader/+/40788
where .c files were renamed to .cpp to avoid a warning treated as error:
command line option ‘-fno-operator-names’ is valid for C++/ObjC++ but not for C [-Werror]

Instead, explicitly add the flag to compile with warnings-as-errors
(when enabled) for code that is actively maintained by us, and make
Subzero's legacy LLVM dependencies a separate library compiled without
warnings-as-errors.

Also revert CMakeSettings.json changes that were inadvertently committed.

Bug: b/141361933
Bug: chromium:1059196
Change-Id: I27dd57d954a413a710613b0bfa2e69bd5242853a