GroupOps: Don't consider helper invocations as active
Helper invocations (those used to calculate derivitives) have an
activeLaneMask(), but an inactive storesAndAtomicsMask().
Document these two functions, and add activeStoresAndAtomicsMask() which
is a simple bitwise-and of these two (but with proper documenation).

Switch group ops to use this new function to ignore helper invocations.
The spec is not clear on whether this is more 'correct', but as b/151137030
suggests, the old behaviour is somewhat less intuitive.

Bug: b/151137030
Change-Id: Id3c88012b5b63e0056d9ca18b9dde22e1506cb68