Allow 3D filtering in the Blitter
According tot he Vulkan spec, about vkCmdBlitImage:
"As the sizes of the source and destination extents can
 differ in any dimension, texels in the source extent
 are scaled and filtered to the destination extent."

Logic was added to allow filtering in 3 dimensions in
the Blitter. Since it requires more instructions, a
state member was added in order to only filter in Z
whenever we actually to do so (when the number of
slices differ between source and destination regions).

Fixes all tests mentioned in the issue.

Bug: b/150155499
Change-Id: I63da92db94d8d0394a06271cd47d4930cf5f726a