Eliminate LRUSnapshotCache<>::add() virtual method overhead
By using protected inheritance, we ensure that LRUSnapshotCache<> can't
be upcasted to its base class LRUCache<>. Hence we can't call the base
class's add() method expecting the derived class's overloaded add()
method to be called. Thus it no longer has to be virtual.

This disallows treating an LRUSnapshotCache<> as a LRUCache<>, but
that's not currently needed nor is it expected to be needed in the
future.

Bug: b/137649247
Change-Id: I3e1a870bb708575edb0e72f32603093baf1bebc0