WIP-sampler-state-id-keep-alive
Increment the sampler ID count for both Vulkan sampler objects and
routines using this ID which are still in the sampling routine cache.

This ensures that if an app destroys all samplers with a certain state,
and then creates a new one with the same state, we can still reuse the
sampling routine that is still in the cache for this state. This may
occur for objects going in and out of view when the app aggressively
destroys and recreates the Vulkan sampler objects for them.

Change-Id: I0ca28c592711c83796f4b75dbb16ff60974fff47