Identify image views based on their state
It is typical for apps to have many image views (and buffer views) with
the same parameters. This change avoids generating new sampling routines
for image views with identical state.

The vk::Identifier class compresses the state into a 32-bit value that
is used as part of the key for the sampling routine cache.

Bug: b/151235334
Change-Id: I00fc19a91639803273d0f7d6b9fed9fc5b481898