Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > I understand why you're making this change, but it somewhat goes counter to how CMake is typically used. Normally, with multiple configurations, you must set the variables separately. This might be a bit painful at first, but it gives flexibility.
> > 
> > With this change, if you happen to have the PowerVR folder in your path, you cannot set SWIFTSHADER_BUILD_PVR to FALSE to force it not to build for the current configuration. This variable effectively becomes meaningless once the folder is present (i.e. once the submodule is checked out).
> > 
> > IMHO, I see two ways to go about this:
> > 
> > 1) Revert this change, and leave it as it was, so that we can control which CMake configs will build PVR or not.
> > 
> > 2) Remove SWIFTSHADER_BUILD_PVR as an option (cache variable), and have CMake auto-add the targets if PVR is present, as you're doing, and the user must 'git submodule update --init third_part/PowerVR_Examples' manually to get PowerVR, and it will automatically appear in all configs.
> > 
> > We could add a non-cache variable to perform step #2 in CMakeLists as a convenience (SWIFTSHADER_GET_PVR?).
> 
> Thanks for raising that concern and offering a solution. I like that last idea except it wouldn't allow me to set the switch for obtaining the PowerVR samples from within Visual Studio's CMake cache editor. If we do make it a cache variable, we've come full circle except for the name change. :)
> 
> Maybe that is all that's needed here, to avoid confusion with other options that have an effect when they get disabled again?

I think we shouldn't use CMake to do the submodule update --init for us using a cache variable. Rather, we could add a script (batch, sh, python?) to run the command for us, if that makes it easier. But most people know how to update a submodule using git, or use a GUI that can do it for them.

One thought I had is that maybe we could hackily use a CMake cache var, then check if it's true, do the submodulate update --init, and set it to false. Not a huge fan, but I think it would work.

Patch-set: 2
Reviewer: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>