Fix asserts due to missing texture formats
In has16bitTextureFormat(), we were missing several formats. Also this
function was poorly named, as it made it seem like we were just
duplicating the work of the bytes() function but giving less
informations. What has16bitTextureFormat() is really checking for is if
the texture format is 16 bit Packed. So it's been renamed to
has16bitPackedTextureFormat()

Bug: b/150076646
Change-Id: I8082e8e9d3845cfa3b5c3ec4a269454457d558ce