Update patch set 4
Patch Set 4:

> Patch Set 3:
> 
> I still argue that the mutex should not be public at all, and that there should be a single atomic get_or_create() function that takes a template-function for the create behaviour.

It's not uncommon for methods to have preconditions, which this change both spells out and validates now. So I still think it's a significant improvement.

Anyway, one way to make it not public is to have an access guard which holds the lock: https://swiftshader-review.googlesource.com/c/SwiftShader/+/42868 (only hides it outside of Device, could be made tighter obviously). Making the mutex locking implicit has downsides though: One could inadvertently be locking and unlocking the mutex twice, once for the query and once for adding the newly generated routine. That would cause generating the same routine multiple times. With an explicit lock_guard<> it is made clearer that the locking is scope-based.

Patch-set: 4
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>