CMake: Emit coverage-toolchain.txt file next to ICD json
When `SWIFTSHADER_EMIT_COVERAGE` is enabled, generate a new `coverage-toolchain.txt` text file next to the `vk_swiftshader_icd.json` file. This file contains a path to the C++ compiler toolchain used to build swiftshader, so that the regres tooling can locate the llvm tools used to parse the coverage data.

Bug: b/152192800
Change-Id: I3aa718206c3262bcd05e13e854002dea8f40801f