Regres: Add support for generating collated coverage info
`run_testlist` now supports the `--coverage` flag, which will collate the coverage information from each and every deqp test that's run.
This information is written to a `coverage.json` file which can be consumed by a html browser.

An early version of this can be found here (likely to move):
https://ben-clayton.github.io/swiftshader-coverage/

Bug: b/152192800
Change-Id: I52434f1ce30e6a091d2932fbae309cd81809cb79