Coverage: Add 'turbo-cov': a faster reimplementation of llvm-cov
`llvm-cov` can either emit data in json or lcov formats, where json is the faster of the two.
`llvm-cov`'s output is directly piped into regres, where the json is immediately deserialized again.
The cost of serializing and deserializing is surprisingly high.

This change replaces the use of `llvm-cov` with `turbo-cov`, which simply emits a binary stream and offers up to 3x speed improvement, dramatically lowering the time taken to produce coverage for a full deqp test run.

Bug: b/152339534
Change-Id: I9292f3c27e016cf508557edf4da7656db81c2b07