Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > I think the opposite should be happening, i.e., all parameters should be accessed through the ImageView. This is extremely error prone.
> > 
> > Could you elaborate? These were already accessible with getter methods. I'm not sure what the "opposite" would look like exactly.
> > 
> > While we do want to be careful not to use the image's format when we should be using the view's format, there's ample test coverage for that, and there's not that many other cases like that.
> > 
> > If anything, it should be on the ImageView class to not expose the Image directly. Image itself has nothing to hide in a reasonable manner.
> 
> Well, for existing getters without arguments, it probably doesn't change anything, except making it less explicit when you read the code that the value is constant and can't be modified.

I'm not sure that's relevant when reading the code. To take an example directly from this patch:

"image->getMipLevels() - range.baseMipLevel" became "image->mipLevels - range.baseMipLevel"

Which of these fields can be modified, and is that relevant? Actually range.baseMipLevel can be modified, because range was passed by value (we might want to fix that). image->mipLevels can't be modified, both because the image pointer is const, and the field itself is const.

I think the distinction becomes important only when writing code, and the compiler will complain loudly when attempting to modify the image's fields.

Instead what's important while reading the code is that getMipLevels() could involve some calculation and even change for the same object, while mipLevels is clearly just a data field, and is immutable. In other words, getMipLevels() allows for "modification".

Getter methods are very valuable when we want a level of abstraction that allows for changing the implementation independently of the interface, but that's not the case here. These are the original creation parameters for the image.

> For other parameters, like device, flags, extent and tiling, I don't think they should be exposed. Either they are internal Image configuration or, in the case of extent, could easily be misused by the ImageView object, where the getter is very explicit that you need to pass proper parameters to get the correct value.

Aside from device, I beg to differ. These parameters are all important to correctly interpret the raw memory that the image maps onto. Unless our shaders were to call abstract getTexel() methods to access that memory, attempts to encapsulate this seem futile to me.

Patch-set: 1
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>