Add script for updating llvm 10
llvm now lives in a mono-repo along with clang, libc, lldb and a whole bunch
of other projects (~90k files at time of writing).
SwiftShader only requires the llvm project from this repo, and does not wish
to pull in everything else.

The script performs a sparse checkout of the llvm project, and copies it to
the third_party directory, replacing all the files. If there are any local
modifications made to the directory in third_party, these are re-applied as a
separate change.

Bug: b/152339534
Change-Id: I75a7bf9e3f4deb7cbf6ff3f7f9428f07b8422800