CMake: Don't export LLVM_INCLUDE_DIR to PARENT_SCOPE
The base change moved the LLVM rules out to a separate CMakeLists.txt file, while trying to keep the behaviour as close as possible to original. `LLVM_INCLUDE_DIR` is not needed to be exposed if we made it a public include directory on the llvm target, so do that.

Bug: b/145758253
Bug: b/152339534
Change-Id: I6e409da409d9a28847ac5bb4074899d70bb38145