Update patch set 2
Patch Set 2:

> I think we shouldn't use CMake to do the submodule update --init for us using a cache variable. Rather, we could add a script (batch, sh, python?) to run the command for us, if that makes it easier. But most people know how to update a submodule using git, or use a GUI that can do it for them.

I'd just like to widen the options for obtaining the samples, without resorting to additional scripts or command line use. If some team members prefer the latter, that's fine, but I prefer having control over it from the GUI. Visual Studio doesn't come with great Git support, but does offer a nice interface for editing the CMake cache.

I don't think there's anything wrong with using CMake to do the submodule checkout. It works on all platforms. And even if it didn't, this is just a by-developers-for-developers convenience thing. Are there any significant downsides even?

> One thought I had is that maybe we could hackily use a CMake cache var, then check if it's true, do the submodulate update --init, and set it to false. Not a huge fan, but I think it would work.

That suits my needs. It's just not clear to me why you consider that any better. Once the submodule is downloaded, it's always there, unless one forcibly deletes it (can't see why anyone would do that). So I don't see any harm in keeping the var true to reflect that.

Anyway, I've lost track of precisely what changes would be needed here that satisfy both our wishes. Any chance it can be landed as-is and then you refine it as part of the CMake overhaul? I've had another need for this while trying to test different configurations of multiple checkouts and it's super-annoying to enable the samples on each one over and over.

Patch-set: 2
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>