Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > I think we shouldn't use CMake to do the submodule update --init for us using a cache variable. Rather, we could add a script (batch, sh, python?) to run the command for us, if that makes it easier. But most people know how to update a submodule using git, or use a GUI that can do it for them.
> 
> I'd just like to widen the options for obtaining the samples, without resorting to additional scripts or command line use. If some team members prefer the latter, that's fine, but I prefer having control over it from the GUI. Visual Studio doesn't come with great Git support, but does offer a nice interface for editing the CMake cache.
> 
> I don't think there's anything wrong with using CMake to do the submodule checkout. It works on all platforms. And even if it didn't, this is just a by-developers-for-developers convenience thing. Are there any significant downsides even?

The thing is, most options are for configuring the build, so it's confusing to mix that with developer options. However, it's fine, but what I find worse is that this variable will only work the first time. Let's say someone else updates the submodule to a later version of PowerVR, and you already used this variable to get it, then you won't be able to update it using this variable.

To fix this, reset the variable back to false as soon as you perform the git submodule update --init, AND remove the check to see if the PowerVR directory exists. This way, you can set the variable, generate, and it will perform the git submodule update --init.

An alternative is to keep the variable as is, but _always_ perform the git submodule update --init if the variable is true (don't check if directory exists). This will be a no-op, unless it ever gets updated or deinited.


> > One thought I had is that maybe we could hackily use a CMake cache var, then check if it's true, do the submodulate update --init, and set it to false. Not a huge fan, but I think it would work.
> 
> That suits my needs. It's just not clear to me why you consider that any better. Once the submodule is downloaded, it's always there, unless one forcibly deletes it (can't see why anyone would do that). So I don't see any harm in keeping the var true to reflect that.

As explained above, although deleting it (via git submodule deinit) is unlikely, having the submodule get updated in the repo is likely to occur.


> Anyway, I've lost track of precisely what changes would be needed here that satisfy both our wishes. Any chance it can be landed as-is and then you refine it as part of the CMake overhaul? I've had another need for this while trying to test different configurations of multiple checkouts and it's super-annoying to enable the samples on each one over and over.

I understand, but what I'm asking is a small incremental change to what you have here. Would you mind trying it?

Patch-set: 2
Reviewer: Gerrit User 9669 <9669@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>