Reactor: Disable more warnings to get LLVM 10 building
There's yet more unnecessary semicolons knocking around in the public LLVM headers, so expand the use of the `diagnostic ignored` directive. I'll try and upstream some fixes for this.

`LegacyIRCompileLayer` is now annotated with deprecated, so we need to disable the warning for that too (and move to the new JIT in time).

Bug: b/152339534
Change-Id: Id84fe88fd2f6b50000e37a8b84f52742ca9db646