Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > This CL breaks SwANGLE, at least angle_deqp_khr_gles3_tests:
> > dEQP.KHR_GLES3/packed_depth_stencil_blit_depth24_stencil8
> > dEQP.KHR_GLES3/packed_depth_stencil_blit_depth32f_stencil8
> > https://luci-milo.appspot.com/p/chromium/builders/ci/win-swangle-tot-swiftshader-x64/4741
> > 
> > The regression range has 5 CLs,
> > https://swiftshader.googlesource.com/SwiftShader/+log/088add4e7b11daa57a2a4d04b6b230a85739e22b..77090261dabcf44c7b5a0f97bdf35e812a13a66a
> > so I've bisected it to this one.
> 
> Confirmed that this CL is also responsible for the angle_deqp_gles3_tests failures (in the same build):
> dEQP.GLES3/functional_fbo_blit_depth_stencil_depth24_stencil8_basic
> dEQP.GLES3/functional_fbo_blit_depth_stencil_depth24_stencil8_scale
> dEQP.GLES3/functional_fbo_blit_depth_stencil_depth24_stencil8_stencil_only
> dEQP.GLES3/functional_fbo_blit_depth_stencil_depth32f_stencil8_basic
> dEQP.GLES3/functional_fbo_blit_depth_stencil_depth32f_stencil8_scale
> dEQP.GLES3/functional_fbo_invalidate_sub_unbind_blit_color
> dEQP.GLES3/functional_fbo_invalidate_sub_unbind_blit_depth
> dEQP.GLES3/functional_fbo_invalidate_whole_unbind_blit_color
> dEQP.GLES3/functional_fbo_invalidate_whole_unbind_blit_depth
> dEQP.GLES3/functional_fbo_msaa_2_samples_depth24_stencil8
> dEQP.GLES3/functional_fbo_msaa_2_samples_depth32f_stencil8
> dEQP.GLES3/functional_fbo_msaa_4_samples_depth24_stencil8
> dEQP.GLES3/functional_fbo_msaa_4_samples_depth32f_stencil8
> 
> Looks like Identifier doesn't handle depth/stencil formats properly.

Thanks a ton for narrowing it down for us!

Yes, VkImageSubresourceRange::aspectMask wasn't being taken into account. Should be an easy fix...

> Not sure what are the practices in SwiftShader group.
> Would you have minded if I've reverted this?

In general, unless it's already affecting users or multiple developers, I think it's preferable to hold off on reverting things until we have a clearer picture of whether we can do a fix-forward or if that's too risky. We usually have a few hours or maybe even a full day to assess that. Reverts are not without unpleasant consequences so I think some thought should go into the decision. I'm also of the opinion that postsubmit tests should either be able to cope with some level of temporary regressions, or we move them to presubmit if that's not acceptable.

In this case I think a fix-forward will be perfectly okay.

Patch-set: 11