Update patch set 5
Patch Set 5: Verified+1

> Patch Set 4: Code-Review-1
> 
> I would much rather see https://swiftshader-review.googlesource.com/c/SwiftShader/+/42869 land than this.

Definitely, but I'd still like to land this as an incremental improvement.

> Exposing internal mutexes publicly for synchronization is horribly error prone, which is clearly why you've created this change. So let's not do that, and just make a clean, non-error-prone API.

I don't disagree about the cleaner API, but using a mutex externally isn't necessarily that error prone. I could put the routine creation in a separate function and SpirvShader::getImageSampler() would be much smaller and almost no different from the new getOrCreate() method, having locking semantics that is easy to follow. It's only if we were passing the mutex around in non-trivial ways that things would get error-prone, but that's true for refactorings or extensions of getOrCreate() as well. It's easier to ensure things don't get out of hand when the mutex is internal though, which would be more important for a less straightforward class.

The callback isn't necessarily inlined, and could have a bit of overhead. This use case isn't that performance critical, but if it was I wouldn't shy away from exposing the mutex. This specific case is strongly in favor of getOrCreate(), but it's not always that black-and-white.

> try_lock_unrecoverable() looks crazy slow, and is a neon sign that the design is _very wrong_.

I'm not sure how that follows. But first of all, it's not really that slow. There was only one timeout, and that test was already close to exceeding it. That was due to regenerating sampling routines too often, which is now fixed and this change no longer causes timeouts.

If the getOrCreate() implementation was more complex and called other methods which require the lock to be held, I would have used ASSERT_LOCKED() in that change as well. It's unfortunate that std::try_lock() requires being called on a separate thread, but it's not like other asserts don't perform some non-trivial computation to verify something. Note that I could have made the thread persistent and/or detach it to make this faster, but that comes with other downsides.

> As Antonio pointed out, synchronization can be verified with static analysis, either with microsoft's SAL annotations or clang's offerings. But still, we shouldn't be expecting mutexes to be locked outside of the class.

Until it's been standardized, I'd rather rely on something that's under our control.

> In the very unlikely case that the above solutions aren't doable, as a last resort, you can make the functions that require the lock to be held to take a std::unique_lock<> parameter, or some sub-class that wraps the std::unique_lock<>. This will force the caller to acquire the lock before calling.

Note that unique_lock<> does not guarantee that the mutex is already locked.

I did the wrapper thing in https://swiftshader-review.googlesource.com/c/SwiftShader/+/42868, but obviously https://swiftshader-review.googlesource.com/c/SwiftShader/+/42869 is superior.

I'll hold off landing this until 42869 is ready to land, if you prefer that.

Patch-set: 5
Reviewer: Gerrit User 5005 <5005@aea13c4a-0b89-3eca-aee9-e193b1b77aa4>
Label: Verified=+1