Fix completeness test of OpenGL ES immutable textures
Immutable textures created with glTexStorage*() are always mipmap and
sampler complete. The spec states that

"Array levels k where k < levelbase or k > q arebinsignificant to the
definition ofcompleteness."

Where q is no larger than levelmax, which is in turn limited to
levelimmut - 1 which was specified at glTexStorage.

Change-Id: Ieaa6be856fd35cf2827e753d7c777ebf784027ef
Tests: dEQP-GLES*
Fixes: b/152740217