Regres: Further optimizations for coverage
Compress common spans into groups and index these.
Reduces the full coverage json from ~500MB to just under 100MB.

Use zlib compression on the final file instead of a zip. The compression algorithm is the same, and produces roughly the same sized output, but a raw zlib file is much quicker to decompress with pako instead of jszip.

Bug: b/152192800
Change-Id: Id444ddf4027b6ace03570719f159ae14e481cd37