Regres: Add invertCommon optimization.
invertCommon looks for tree nodes with the majority of the child nodes with the same spans.
This span is promoted up to the parent, and the children have the
span inverted.

This causes the span IDs to flip (inclusive / exlusive) each time they're seen while decending the tree. This adds a bit of complexity to the parsing of the data, but the file size reduction is significant.

Bug: b/152192800
Change-Id: I5ec294d42004d936004c27ef15bf94a7143372ba